fix: selectedIds revert to initial value of null
React setState functions are batched together so
there was no control over the timing of when the
data gets deleted. The data would be deleted first
and the id would try to access undefined data.
This means that experienceData[0].id would not
exist and there was nothing to read from, but that
issue doesn't happen for the dynamic id selections
and deletions that happen afterward.

On the rest of the dynamic id selections, they are
direct string values rather than getting the id
from a variable so there is no undefined variable
getting accessed. This makes it so that the order
does not matter for id selection or deletion.